Fixed warnings C4018 and C4244.

+++ I18NTest_vs2012/I18NTest/i18n/Entry.cpp
@@ -80,7 +80,7 @@ std::string TranslateEntry::toString()
     auto str = Entry::toString();
     
     if (!_translations.empty()) {
-        for (int i = 0; i < _translations.size(); i++) {
+        for (size_t i = 0; i < _translations.size(); i++) {
             str += "\n\t";
             str += " Translation";
             if (1 < _translations.size()) {
